Migrate and enhance `filter_out_by_uuid` tool.
Moved `filter_out_by_uuid` from `DD_tools` to the `TreeOfLife_toolbox` package. Improved documentation, streamlined class methods, and clarified configurations for better usability. Added attributes and detailed explanations for filters, schedulers, and runners.
