--- src/DD_tools/filter_out_by_uuid/README.md
@@ -1,6 +1,0 @@
-tool to filter out using table of Tree of life `uuid`s
-
-Additional config fields:
-
-* `uuid_table_path` - path to uuid table to filter out
-* `look_up_table_path` - path to look up table with `uuid - file_name` information
